Split 'enclaver run-eif' into its own 'encalver-run' binary
