+++ .github/workflows/release.yaml
@@ -26,6 +26,7 @@ jobs:
           name: ${{ matrix.build-target }}
           path: |
             enclaver/target/${{ matrix.build-target }}/release/enclaver
+            enclaver/target/${{ matrix.build-target }}/release/enclaver-run
             enclaver/target/${{ matrix.build-target }}/release/odyn
 
   publish-images:
@@ -50,7 +51,7 @@ jobs:
         run: |
           mv x86_64-unknown-linux-musl amd64
           mv aarch64-unknown-linux-musl arm64
-          chmod 755 amd64/odyn amd64/enclaver arm64/odyn arm64/enclaver
+          chmod 755 amd64/odyn amd64/enclaver-run arm64/odyn arm64/enclaver-run
 
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v2
